Chore: (BRD-32) domain-member: domain-email 모듈 통합
