+++ board-system-application/application-member/build.gradle.kts
@@ -1,6 +1,5 @@
 dependencies {
     implementation(project(":board-system-application:application-core"))
     implementation(project(":board-system-domain:domain-member"))
-    implementation(project(":board-system-domain:domain-email"))
     implementation(project(":board-system-domain:domain-core"))
 }
