--- board-system-domain/domain-email/build.gradle.kts
@@ -1,3 +1,0 @@
-dependencies {
-    implementation(project(":board-system-domain:domain-core"))
-}
