--- board-system-domain/domain-email/src/main/kotlin/com/ttasjwi/board/system/email/domain/event/EmailVerificationStartedEvent.kt
@@ -1,22 +1,0 @@
-package com.ttasjwi.board.system.email.domain.event
-
-import com.ttasjwi.board.system.core.domain.event.DomainEvent
-import java.time.ZonedDateTime
-
-class EmailVerificationStartedEvent(
-    email: String,
-    code: String,
-    codeCreatedAt: ZonedDateTime,
-    codeExpiresAt: ZonedDateTime,
-) : DomainEvent<EmailVerificationStartedEvent.Data>(
-    occurredAt = codeCreatedAt,
-    data = Data(email, code, codeCreatedAt, codeExpiresAt)
-) {
-
-    class Data(
-        val email: String,
-        val code: String,
-        val codeCreatedAt: ZonedDateTime,
-        val codeExpiresAt: ZonedDateTime,
-    )
-}
