--- board-system-domain/domain-email/src/main/kotlin/com/ttasjwi/board/system/email/domain/service/EmailSender.kt
@@ -1,12 +1,0 @@
-package com.ttasjwi.board.system.email.domain.service
-
-import com.ttasjwi.board.system.member.domain.model.Email
-
-interface EmailSender {
-
-    fun send(
-        address: Email,
-        subject: String,
-        content: String
-    ): Result<Unit>
-}
