--- board-system-domain/domain-email/src/main/kotlin/com/ttasjwi/board/system/email/domain/service/EmailVerificationAppender.kt
@@ -1,11 +1,0 @@
-package com.ttasjwi.board.system.email.domain.service
-
-import com.ttasjwi.board.system.member.domain.model.Email
-import com.ttasjwi.board.system.email.domain.model.EmailVerification
-import java.time.ZonedDateTime
-
-interface EmailVerificationAppender {
-
-    fun append(emailVerification: EmailVerification, expiresAt: ZonedDateTime)
-    fun removeByEmail(email: Email)
-}
