--- board-system-domain/domain-email/src/main/kotlin/com/ttasjwi/board/system/email/domain/service/EmailVerificationEventCreator.kt
@@ -1,11 +1,0 @@
-package com.ttasjwi.board.system.email.domain.service
-
-import com.ttasjwi.board.system.email.domain.event.EmailVerificationStartedEvent
-import com.ttasjwi.board.system.email.domain.event.EmailVerifiedEvent
-import com.ttasjwi.board.system.email.domain.model.EmailVerification
-
-interface EmailVerificationEventCreator {
-
-    fun onVerificationStarted(emailVerification: EmailVerification): EmailVerificationStartedEvent
-    fun onVerified(emailVerification: EmailVerification): EmailVerifiedEvent
-}
