--- board-system-domain/domain-email/src/main/kotlin/com/ttasjwi/board/system/email/domain/service/EmailVerificationFinder.kt
@@ -1,9 +1,0 @@
-package com.ttasjwi.board.system.email.domain.service
-
-import com.ttasjwi.board.system.member.domain.model.Email
-import com.ttasjwi.board.system.email.domain.model.EmailVerification
-
-interface EmailVerificationFinder {
-
-    fun findByEmailOrNull(email: Email): EmailVerification?
-}
