+++ board-system-domain/domain-member/src/main/kotlin/com/ttasjwi/board/system/member/domain/service/EmailSender.kt
@@ -1,0 +1,12 @@
+package com.ttasjwi.board.system.member.domain.service
+
+import com.ttasjwi.board.system.member.domain.model.Email
+
+interface EmailSender {
+
+    fun send(
+        address: Email,
+        subject: String,
+        content: String
+    ): Result<Unit>
+}
