+++ board-system-domain/domain-member/src/main/kotlin/com/ttasjwi/board/system/member/domain/service/EmailVerificationFinder.kt
@@ -1,0 +1,8 @@
+package com.ttasjwi.board.system.member.domain.service
+
+import com.ttasjwi.board.system.member.domain.model.Email
+import com.ttasjwi.board.system.member.domain.model.EmailVerification
+
+interface EmailVerificationFinder {
+    fun findByEmailOrNull(email: Email): EmailVerification?
+}
