--- board-system-domain/domain-member/src/test/kotlin/com/ttasjwi/board/system/email/domain/model/fixture/EmailFixture.kt
@@ -1,9 +1,0 @@
-package com.ttasjwi.board.system.email.domain.model.fixture
-
-import com.ttasjwi.board.system.member.domain.model.Email
-
-fun emailFixture(
-    value: String = "test@gmail.com"
-): Email {
-    return Email(value)
-}
