+++ board-system-domain/domain-member/src/test/kotlin/com/ttasjwi/board/system/member/domain/event/fixture/EmailVerifiedEventFixtureTest.kt
@@ -1,0 +1,42 @@
+package com.ttasjwi.board.system.member.domain.event.fixture
+
+import com.ttasjwi.board.system.core.time.fixture.timeFixture
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.DisplayName
+import org.junit.jupiter.api.Test
+
+@DisplayName("EmailVerifiedEventFixture 테스트")
+class EmailVerifiedEventFixtureTest {
+
+    @Test
+    @DisplayName("인자를 전달하지 않아도 기본값들을 가지고 있다.")
+    fun test1() {
+        val event = emailVerifiedEventFixture()
+        val data = event.data
+
+        assertThat(event.occurredAt).isNotNull()
+        assertThat(data.email).isNotNull()
+        assertThat(data.verifiedAt).isNotNull()
+        assertThat(data.verificationExpiresAt).isNotNull()
+    }
+
+    @Test
+    @DisplayName("커스텀하게 필드를 지정할 수 있다.")
+    fun test2() {
+        val email = "fire@gmail.com"
+        val verifiedAt = timeFixture(minute = 11)
+        val verificationExpiresAt = timeFixture(minute = 41)
+
+        val event = emailVerifiedEventFixture(
+            email = email,
+            verifiedAt = verifiedAt,
+            verificationExpiresAt = verificationExpiresAt
+        )
+        val data = event.data
+
+        assertThat(event.occurredAt).isEqualTo(verifiedAt)
+        assertThat(data.email).isEqualTo(email)
+        assertThat(data.verifiedAt).isEqualTo(verifiedAt)
+        assertThat(data.verificationExpiresAt).isEqualTo(verificationExpiresAt)
+    }
+}
