--- board-system-external/external-email-sender/src/main/kotlin/com/ttasjwi/board/system/email/domain/external/EmailSenderImpl.kt
@@ -1,13 +1,0 @@
-package com.ttasjwi.board.system.email.domain.external
-
-import com.ttasjwi.board.system.core.annotation.component.AppComponent
-import com.ttasjwi.board.system.member.domain.model.Email
-import com.ttasjwi.board.system.email.domain.service.EmailSender
-
-@AppComponent
-class EmailSenderImpl : EmailSender {
-
-    override fun send(address: Email, subject: String, content: String): Result<Unit> {
-        TODO("Not yet implemented")
-    }
-}
