+++ board-system-external/external-redis/build.gradle.kts
@@ -1,4 +1,4 @@
 dependencies {
     implementation(project(":board-system-domain:domain-core"))
-    implementation(project(":board-system-domain:domain-email"))
+    implementation(project(":board-system-domain:domain-member"))
 }
